Refactor GlobalSideNav --> AppSideNavigation
