+++ src/app/base/components/AppSideNavigation/NavigationBanner/NavigationBanner.test.tsx
