+++ src/app/base/components/AppSideNavigation/NavigationBanner/NavigationBanner.tsx
