+++ src/app/base/components/AppSideNavigation/NavigationBanner/index.ts
