+++ src/app/base/components/AppSideNavigation/_index.scss
@@ -1,4 +1,4 @@
-@mixin GlobalSideNav {
+@mixin AppSideNavigation {
   .l-navigation .p-navigation__logo-title {
     margin-left: $spv--large;
   }
