+++ src/app/base/components/AppSideNavigation/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AppSideNavigation";
