+++ src/app/base/components/AppSideNavigation/utils.ts
