--- src/app/base/components/GlobalSideNav/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./GlobalSideNav";
